Refactor: modularized api controllers, modularized rtk endpoint
