+++ api/src/controllers/planControllers/getMpSubscriptionPlanIdApi.ts
@@ -1,0 +1,31 @@
+import { Request, Response } from "express";
+import { getPlanDetail } from "../../services/plans/PlanServiceAPI";
+
+
+
+
+export const getMpSubscriptionPlanIdApi = async (req: Request, res: Response) => {
+	// * Funcion para obtener el id del plan
+	const { id } = req.params;
+
+	try {
+		const getId = await getPlanDetail(id);
+		if (getId) {
+			res.status(200).json({
+				msg: "Id del plan obtenido correctamente",
+				id: getId,
+			});
+		} else {
+			res.status(500).json({
+				error: true,
+				msg: "Hubo un error al obtener el id del plan",
+			});
+		}
+	} catch (err) {
+		console.log(err);
+		res.status(500).json({
+			error: true,
+			msg: "Hubo un error al obtener el id del plan",
+		});
+	}
+};
