+++ client/package.json
@@ -41,6 +41,7 @@
     "react-router-dom": "^6.10.0",
     "react-social-media-embed": "^2.3.5",
     "react-timeago": "^7.1.0",
+    "redux-persist": "^6.0.0",
     "sweetalert2": "^11.7.3",
     "sweetalert2-react-content": "^5.0.7",
     "swiper": "^9.2.4",
